Configure linters for json, md, scss
This commit adds linters for markdown through remark and scss through
stylelinter. The configuration of each has been updated to work with eslint,
and eslint has been configured to apply the its rules to json files. Files with
the mdx extension are purposely ignored by eslint, but will be processed by
the remark linter.

It creates a new scripts directory to call each linter with the appropriate
options and flags. This commit also updates the required node version to the
latest LTS which is 16.13.0.